chore(): Refactor types
