+++ src/components/Extension/CalDotComExtension/types.ts
@@ -1,3 +1,6 @@
+import { ComponentProps } from 'react'
+import { CalDotComScheduling } from '@awell_health/ui-library'
+
 export enum ActionKey {
   BOOK_APPOINTMENT = 'bookAppointment',
 }
@@ -5,3 +8,7 @@ export enum ActionKey {
 export type BookAppointmentFields = {
   calLink: string
 }
+
+export type BookingSuccessfulFunction = ComponentProps<
+  typeof CalDotComScheduling
+>['onBookingSuccessful']
